# adapter/src/main/java/eu/cokeman/velomarker/openapi/api/DraftsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.22.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package eu.cokeman.velomarker.openapi.api;

import eu.cokeman.velomarker.openapi.model.RouteDraftGroupGeometriesResponseDto;
import eu.cokeman.velomarker.openapi.model.RouteDraftRequestDto;
import eu.cokeman.velomarker.openapi.model.RouteDraftResponseDto;
import eu.cokeman.velomarker.openapi.model.RouteDraftSummaryDto;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.22.0")
@Validated
@Controller
@Tag(name = "Drafts", description = "the Drafts API")
public interface DraftsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_ROUTE_DRAFT = "/route/drafts";
    /**
     * POST /route/drafts : Create a new draft
     *
     * @param routeDraftRequestDto  (required)
     * @return Draft created (status code 201)
     *         or Limit exceeded (max 50 per user) or invalid data (status code 400)
     *         or Unauthorized (status code 401)
     *         or Name already in use by this user (status code 409)
     */
    @Operation(
        operationId = "createRouteDraft",
        summary = "Create a new draft",
        tags = { "Drafts" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Draft created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RouteDraftResponseDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Limit exceeded (max 50 per user) or invalid data"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "409", description = "Name already in use by this user")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = DraftsApi.PATH_CREATE_ROUTE_DRAFT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<RouteDraftResponseDto> createRouteDraft(
        @Parameter(name = "RouteDraftRequestDto", description = "", required = true) @Valid @RequestBody RouteDraftRequestDto routeDraftRequestDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"profile\" : \"profile\", \"groupId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"distanceKm\" : 0.8008281904610115, \"elevationGain\" : 6, \"waypoints\" : \"waypoints\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"groupName\" : \"groupName\", \"name\" : \"name\", \"elevationLoss\" : 1, \"dayNumber\" : 5, \"geometry\" : { \"coordinates\" : [ [ 0.8008281904610115, 0.8008281904610115, 0.8008281904610115 ], [ 0.8008281904610115, 0.8008281904610115, 0.8008281904610115 ] ], \"type\" : \"LineString\" }, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DELETE_ROUTE_DRAFT = "/route/drafts/{draftId}";
    /**
     * DELETE /route/drafts/{draftId} : Delete draft
     *
     * @param draftId  (required)
     * @return Deleted (status code 204)
     *         or Unauthorized (status code 401)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "deleteRouteDraft",
        summary = "Delete draft",
        tags = { "Drafts" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Deleted"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = DraftsApi.PATH_DELETE_ROUTE_DRAFT
    )
    default ResponseEntity<Void> deleteRouteDraft(
        @Parameter(name = "draftId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("draftId") UUID draftId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_EXPORT_ROUTE_DRAFT_GPX = "/route/drafts/{draftId}/export/gpx";
    /**
     * GET /route/drafts/{draftId}/export/gpx : Export single draft (day) as GPX
     *
     * @param draftId  (required)
     * @return GPX file content (status code 200)
     *         or Unauthorized (status code 401)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "exportRouteDraftGpx",
        summary = "Export single draft (day) as GPX",
        tags = { "Drafts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "GPX file content", content = {
                @Content(mediaType = "application/gpx+xml", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DraftsApi.PATH_EXPORT_ROUTE_DRAFT_GPX,
        produces = { "application/gpx+xml" }
    )
    default ResponseEntity<String> exportRouteDraftGpx(
        @Parameter(name = "draftId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("draftId") UUID draftId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_EXPORT_ROUTE_GROUP_GPX = "/route/drafts/group/{groupId}/export/gpx";
    /**
     * GET /route/drafts/group/{groupId}/export/gpx : Export whole expedition (all days of a group) as GPX
     *
     * @param groupId  (required)
     * @return GPX file content (status code 200)
     *         or Unauthorized (status code 401)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "exportRouteGroupGpx",
        summary = "Export whole expedition (all days of a group) as GPX",
        tags = { "Drafts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "GPX file content", content = {
                @Content(mediaType = "application/gpx+xml", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DraftsApi.PATH_EXPORT_ROUTE_GROUP_GPX,
        produces = { "application/gpx+xml" }
    )
    default ResponseEntity<String> exportRouteGroupGpx(
        @Parameter(name = "groupId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("groupId") UUID groupId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_ROUTE_DRAFT = "/route/drafts/{draftId}";
    /**
     * GET /route/drafts/{draftId} : Get full draft with geometry
     *
     * @param draftId  (required)
     * @return Draft details (status code 200)
     *         or Unauthorized (status code 401)
     *         or Not found or not owned by this user (status code 404)
     */
    @Operation(
        operationId = "getRouteDraft",
        summary = "Get full draft with geometry",
        tags = { "Drafts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Draft details", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RouteDraftResponseDto.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "404", description = "Not found or not owned by this user")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DraftsApi.PATH_GET_ROUTE_DRAFT,
        produces = { "application/json" }
    )
    default ResponseEntity<RouteDraftResponseDto> getRouteDraft(
        @Parameter(name = "draftId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("draftId") UUID draftId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"profile\" : \"profile\", \"groupId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"distanceKm\" : 0.8008281904610115, \"elevationGain\" : 6, \"waypoints\" : \"waypoints\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"groupName\" : \"groupName\", \"name\" : \"name\", \"elevationLoss\" : 1, \"dayNumber\" : 5, \"geometry\" : { \"coordinates\" : [ [ 0.8008281904610115, 0.8008281904610115, 0.8008281904610115 ], [ 0.8008281904610115, 0.8008281904610115, 0.8008281904610115 ] ], \"type\" : \"LineString\" }, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_ROUTE_DRAFT_GROUP_GEOMETRIES = "/route/drafts/group/{groupId}/geometries";
    /**
     * GET /route/drafts/group/{groupId}/geometries : Bulk fetch all day geometries of an expedition (Polyline3DCodec encoded)
     * Zwraca wszystkie geometrie dni grupy w JEDNYM żądaniu, kodowane jako Polyline3DCodec (frontend dekoduje przez decodePolyline3D). Zastępuje N×GET /route/drafts/{id} przy ładowaniu całej wyprawy na mapie (2000+ dni). Sortowane po dayNumber rosnąco. 
     *
     * @param groupId  (required)
     * @return Bulk encoded geometries (status code 200)
     *         or Unauthorized (status code 401)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "getRouteDraftGroupGeometries",
        summary = "Bulk fetch all day geometries of an expedition (Polyline3DCodec encoded)",
        description = "Zwraca wszystkie geometrie dni grupy w JEDNYM żądaniu, kodowane jako Polyline3DCodec (frontend dekoduje przez decodePolyline3D). Zastępuje N×GET /route/drafts/{id} przy ładowaniu całej wyprawy na mapie (2000+ dni). Sortowane po dayNumber rosnąco. ",
        tags = { "Drafts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Bulk encoded geometries", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RouteDraftGroupGeometriesResponseDto.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DraftsApi.PATH_GET_ROUTE_DRAFT_GROUP_GEOMETRIES,
        produces = { "application/json" }
    )
    default ResponseEntity<RouteDraftGroupGeometriesResponseDto> getRouteDraftGroupGeometries(
        @Parameter(name = "groupId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("groupId") UUID groupId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"groupId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"days\" : [ { \"geometryEncoded\" : \"geometryEncoded\", \"dayNumber\" : 0, \"distanceKm\" : 6.027456183070403 }, { \"geometryEncoded\" : \"geometryEncoded\", \"dayNumber\" : 0, \"distanceKm\" : 6.027456183070403 } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LIST_ROUTE_DRAFTS = "/route/drafts";
    /**
     * GET /route/drafts : List authenticated user&#39;s route drafts
     *
     * @return List of drafts (without geometry — call /route/drafts/{id} for full data) (status code 200)
     *         or Unauthorized (status code 401)
     */
    @Operation(
        operationId = "listRouteDrafts",
        summary = "List authenticated user's route drafts",
        tags = { "Drafts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of drafts (without geometry — call /route/drafts/{id} for full data)", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = RouteDraftSummaryDto.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DraftsApi.PATH_LIST_ROUTE_DRAFTS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<RouteDraftSummaryDto>> listRouteDrafts(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"groupName\" : \"groupName\", \"profile\" : \"profile\", \"groupId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"name\" : \"name\", \"elevationLoss\" : 1, \"dayNumber\" : 5, \"distanceKm\" : 0.8008281904610115, \"elevationGain\" : 6, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"groupName\" : \"groupName\", \"profile\" : \"profile\", \"groupId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"name\" : \"name\", \"elevationLoss\" : 1, \"dayNumber\" : 5, \"distanceKm\" : 0.8008281904610115, \"elevationGain\" : 6, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPDATE_ROUTE_DRAFT = "/route/drafts/{draftId}";
    /**
     * PUT /route/drafts/{draftId} : Update draft (name + geometry + metadata)
     *
     * @param draftId  (required)
     * @param routeDraftRequestDto  (required)
     * @return Draft updated (status code 200)
     *         or Unauthorized (status code 401)
     *         or Not found (status code 404)
     *         or Name conflict (status code 409)
     */
    @Operation(
        operationId = "updateRouteDraft",
        summary = "Update draft (name + geometry + metadata)",
        tags = { "Drafts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Draft updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RouteDraftResponseDto.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "404", description = "Not found"),
            @ApiResponse(responseCode = "409", description = "Name conflict")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = DraftsApi.PATH_UPDATE_ROUTE_DRAFT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<RouteDraftResponseDto> updateRouteDraft(
        @Parameter(name = "draftId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("draftId") UUID draftId,
        @Parameter(name = "RouteDraftRequestDto", description = "", required = true) @Valid @RequestBody RouteDraftRequestDto routeDraftRequestDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"profile\" : \"profile\", \"groupId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"distanceKm\" : 0.8008281904610115, \"elevationGain\" : 6, \"waypoints\" : \"waypoints\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"groupName\" : \"groupName\", \"name\" : \"name\", \"elevationLoss\" : 1, \"dayNumber\" : 5, \"geometry\" : { \"coordinates\" : [ [ 0.8008281904610115, 0.8008281904610115, 0.8008281904610115 ], [ 0.8008281904610115, 0.8008281904610115, 0.8008281904610115 ] ], \"type\" : \"LineString\" }, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}