# adapter/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Route Service API
  version: 1.0.0
  description: |
    Route planning: brouter proxy (semaphore-protected), draft persistence,
    elevation profile (Copernicus DEM). Plus admin endpoints to manage BRouter
    segments (.rd5) and profiles.
servers:
  - url: http://localhost:8085
    description: Local development server
paths:
...
  /route/drafts:
    get:
      summary: List authenticated user's route drafts
      operationId: listRouteDrafts
      tags: [Drafts]
      security: []
      responses:
        '200':
          description: List of drafts (without geometry — call /route/drafts/{id} for full data)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteDraftSummary'
        '401':
          description: Unauthorized
    post:
      summary: Create a new draft
      operationId: createRouteDraft
      tags: [Drafts]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteDraftRequest'
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDraftResponse'
        '400':
          description: Limit exceeded (max 50 per user) or invalid data
        '401':
          description: Unauthorized
        '409':
          description: Name already in use by this user
  /route/drafts/{draftId}:
    get:
      summary: Get full draft with geometry
      operationId: getRouteDraft
      tags: [Drafts]
      security: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Draft details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDraftResponse'
        '401':
          description: Unauthorized
        '404':
          description: Not found or not owned by this user
    put:
      summary: Update draft (name + geometry + metadata)
      operationId: updateRouteDraft
      tags: [Drafts]
      security: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteDraftRequest'
      responses:
        '200':
          description: Draft updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDraftResponse'
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '409':
          description: Name conflict
    delete:
      summary: Delete draft
      operationId: deleteRouteDraft
      tags: [Drafts]
      security: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '404':
          description: Not found
  /route/drafts/{draftId}/export/gpx:
    get:
      summary: Export single draft (day) as GPX
      operationId: exportRouteDraftGpx
      tags: [Drafts]
      security: []
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: GPX file content
          content:
            application/gpx+xml:
              schema:
                type: string
        '401':
          description: Unauthorized
        '404':
          description: Not found
  /route/drafts/group/{groupId}/export/gpx:
    get:
      summary: Export whole expedition (all days of a group) as GPX
      operationId: exportRouteGroupGpx
      tags: [Drafts]
      security: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: GPX file content
          content:
            application/gpx+xml:
              schema:
                type: string
        '401':
          description: Unauthorized
        '404':
          description: Not found
  /route/drafts/group/{groupId}/geometries:
    get:
      summary: Bulk fetch all day geometries of an expedition (Polyline3DCodec encoded)
      description: |
        Zwraca wszystkie geometrie dni grupy w JEDNYM żądaniu, kodowane jako Polyline3DCodec
        (frontend dekoduje przez decodePolyline3D). Zastępuje N×GET /route/drafts/{id} przy ładowaniu
        całej wyprawy na mapie (2000+ dni). Sortowane po dayNumber rosnąco.
      operationId: getRouteDraftGroupGeometries
      tags: [Drafts]
      security: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bulk encoded geometries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDraftGroupGeometriesResponse'
        '401':
          description: Unauthorized
        '404':
          description: Not found
...
components:
  schemas:
...
    LineStringGeoJson:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [LineString]
        coordinates:
          type: array
          items:
            type: array
            minItems: 2
            maxItems: 3
            items:
              type: number
              format: double
...
    RouteDraftRequest:
      type: object
      required: [name, geometry, profile]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: User-facing draft name, unique per user
        geometry:
          $ref: '#/components/schemas/LineStringGeoJson'
        profile:
          type: string
          description: BRouter profile used (snapshot at save time)
        distanceKm:
          type: number
          format: double
          nullable: true
        elevationGain:
          type: integer
          nullable: true
        elevationLoss:
          type: integer
          nullable: true
        groupId:
          type: string
          format: uuid
          nullable: true
          description: Wyprawa (grupa) — wspólny identyfikator dni jednej wyprawy
        groupName:
          type: string
          nullable: true
        dayNumber:
          type: integer
          nullable: true
        waypoints:
          type: string
          nullable: true
          description: Waypointy (gminy) jako zakodowana polyline (OPAQUE) — do edycji wczytanego szkicu
    RouteDraftSummary:
      type: object
      required: [id, name, profile, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        profile:
          type: string
        distanceKm:
          type: number
          format: double
          nullable: true
        elevationGain:
          type: integer
          nullable: true
        elevationLoss:
          type: integer
          nullable: true
        groupId:
          type: string
          format: uuid
          nullable: true
        groupName:
          type: string
          nullable: true
        dayNumber:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RouteDraftResponse:
      type: object
      required: [id, name, geometry, profile, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        geometry:
          $ref: '#/components/schemas/LineStringGeoJson'
        profile:
          type: string
        distanceKm:
          type: number
          format: double
          nullable: true
        elevationGain:
          type: integer
          nullable: true
        elevationLoss:
          type: integer
          nullable: true
        groupId:
          type: string
          format: uuid
          nullable: true
        groupName:
          type: string
          nullable: true
        dayNumber:
          type: integer
          nullable: true
        waypoints:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RouteDraftGroupGeometriesResponse:
      type: object
      required: [groupId, days]
      properties:
        groupId:
          type: string
          format: uuid
        days:
          type: array
          items:
            $ref: '#/components/schemas/RouteDraftGroupGeometryItem'
    RouteDraftGroupGeometryItem:
      type: object
      required: [dayNumber, geometryEncoded]
      properties:
        dayNumber:
          type: integer
        geometryEncoded:
          type: string
          description: Polyline3DCodec encoded geometry (lng,lat,z) — frontend decodes via decodePolyline3D
        distanceKm:
          type: number
          format: double
          nullable: true
...